requirements.txt
Django==1.11.10
django-json-widget==0.1.0
django-jsonfield==1.0.1
djangorestframework==3.7.7
mysqlclient==1.3.12
pytz==2018.3
